requirements.txt
fastapi
uvicorn
apscheduler
requests
psycopg2-binary
python-dotenv
